refactor: separate parsing_utils and create file for each parse method
